# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: 112 AI Trainer REST API
  version: 0.3.0
  description: Backend scenario API and Voice Gateway API on separate ports.
servers:
- url: http://127.0.0.1:8000/api/v1
  description: Scenario backend
- url: http://127.0.0.1:8001/api/v1
  description: Voice Gateway
security:
- bearerAuth: []
paths:
...
  /student/scenarios:
    get:
      summary: Scenarios
      operationId: scenarios_api_v1_student_scenarios_get
      parameters:
      - name: x-user-session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /student/sessions:
    get:
      summary: Sessions
      operationId: sessions_api_v1_student_sessions_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: x-user-session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
    post:
      summary: Create
      operationId: create_api_v1_student_sessions_post
      parameters:
      - name: x-user-session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSession'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
...
  /student/classifier:
    get:
      summary: Classifier Catalog
      operationId: classifier_catalog_api_v1_student_classifier_get
      parameters:
      - name: x-user-session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /student/routing/preview:
    post:
      summary: Routing Preview
      operationId: routing_preview_api_v1_student_routing_preview_post
      parameters:
      - name: x-user-session
        in: header
        required: false
        schema:
          type: string
          default: ''
          title: X-User-Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Card'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutingPreview'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    Card:
      properties:
        caller_name:
          type: string
          maxLength: 160
          title: Caller Name
          default: ''
        caller_status:
          type: string
          maxLength: 80
          title: Caller Status
          default: ''
        phone:
          type: string
          maxLength: 40
          title: Phone
          default: ''
        supplied_phone:
          type: string
          maxLength: 40
          title: Supplied Phone
          default: ''
        scene_phone:
          type: string
          maxLength: 40
          title: Scene Phone
          default: ''
        country:
          type: string
          maxLength: 100
          title: Country
          default: Россия
        region:
          type: string
          maxLength: 100
          title: Region
          default: Москва
        city:
          type: string
          maxLength: 100
          title: City
          default: Москва
        district:
          type: string
          maxLength: 100
          title: District
          default: ''
        area:
          type: string
          maxLength: 100
          title: Area
          default: ''
        object:
          type: string
          maxLength: 200
          title: Object
          default: ''
        street:
          type: string
          maxLength: 200
          title: Street
          default: ''
        house:
          type: string
          maxLength: 40
          title: House
          default: ''
        building:
          type: string
          maxLength: 40
          title: Building
          default: ''
        structure:
          type: string
          maxLength: 40
          title: Structure
          default: ''
        apartment:
          type: string
          maxLength: 40
          title: Apartment
          default: ''
        entrance:
          type: string
          maxLength: 40
          title: Entrance
          default: ''
        floor:
          type: string
          maxLength: 40
          title: Floor
          default: ''
        code:
          type: string
          maxLength: 40
          title: Code
          default: ''
        address_note:
          type: string
          maxLength: 500
          title: Address Note
          default: ''
        latitude:
          anyOf:
          - type: number
            maximum: 90.0
            minimum: -90.0
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
            maximum: 180.0
            minimum: -180.0
          - type: 'null'
          title: Longitude
        emergency:
          type: boolean
          title: Emergency
          default: false
        important:
          type: boolean
          title: Important
          default: false
        bookmarked:
          type: boolean
          title: Bookmarked
          default: false
        description:
          type: string
          maxLength: 1999
          title: Description
          default: ''
        incident_type:
          type: string
          maxLength: 1000
          title: Incident Type
          default: ''
        classifier_id:
          type: string
          maxLength: 100
          title: Classifier Id
          default: ''
        classifier_version:
          type: string
          maxLength: 100
          title: Classifier Version
          default: ''
        classifier_group:
          type: string
          maxLength: 100
          title: Classifier Group
          default: ''
        classifier_features:
          items:
            type: string
            maxLength: 1000
          type: array
          maxItems: 3
          title: Classifier Features
        place:
          type: string
          maxLength: 80
          title: Place
          default: ''
        sign:
          type: string
          maxLength: 100
          title: Sign
          default: ''
        detail:
          type: string
          maxLength: 100
          title: Detail
          default: ''
        injured:
          type: boolean
          title: Injured
          default: false
        no_access:
          type: boolean
          title: No Access
          default: false
        threat_to_people:
          type: boolean
          title: Threat To People
          default: false
        offense:
          type: boolean
          title: Offense
          default: false
        injured_offsite:
          type: boolean
          title: Injured Offsite
          default: false
        gasification:
          type: boolean
          title: Gasification
          default: false
        medical_help:
          type: boolean
          title: Medical Help
          default: false
        evacuation:
          type: boolean
          title: Evacuation
          default: false
        fsb_special:
          type: boolean
          title: Fsb Special
          default: false
        traffic_blocked:
          type: boolean
          title: Traffic Blocked
          default: false
        tunnel:
          type: boolean
          title: Tunnel
          default: false
        pedestrian_structure:
          type: boolean
          title: Pedestrian Structure
          default: false
        vehicle_structure:
          type: boolean
          title: Vehicle Structure
          default: false
        telecom_object:
          type: boolean
          title: Telecom Object
          default: false
        construction_site:
          type: boolean
          title: Construction Site
          default: false
        culture_listed_object:
          type: boolean
          title: Culture Listed Object
          default: false
        territorial_oiv:
          type: boolean
          title: Territorial Oiv
          default: false
        territorial_oiv_tinao:
          type: boolean
          title: Territorial Oiv Tinao
          default: false
        territorial_roads_moscow:
          type: boolean
          title: Territorial Roads Moscow
          default: false
        rhbz_polygon:
          type: boolean
          title: Rhbz Polygon
          default: false
        rhbz_moscow:
          type: boolean
          title: Rhbz Moscow
          default: false
        refused:
          type: boolean
          title: Refused
          default: false
        no_contact:
          type: boolean
          title: No Contact
          default: false
        interrupted:
          type: boolean
          title: Interrupted
          default: false
        services:
          items:
            type: string
            maxLength: 160
            minLength: 1
          type: array
          maxItems: 100
          title: Services
      additionalProperties: false
      type: object
      title: Card
    CreateSession:
      properties:
        scenario_id:
          type: string
          title: Scenario Id
        assignment_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Assignment Id
        transport:
          type: string
          enum:
          - text
          - sip
          title: Transport
          default: text
        workstation:
          type: string
          maxLength: 80
          title: Workstation
          default: ''
      type: object
      required:
      - scenario_id
      title: CreateSession
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    RoutingExclusion:
      properties:
        service:
          type: string
          title: Service
        cell:
          type: string
          title: Cell
        reason:
          type: string
          title: Reason
      type: object
      required:
      - service
      - cell
      - reason
      title: RoutingExclusion
    RoutingMapping:
      properties:
        cell:
          type: string
          title: Cell
        incident_type:
          type: string
          title: Incident Type
      type: object
      required:
      - cell
      - incident_type
      title: RoutingMapping
    RoutingPreview:
      properties:
        rules_version:
          type: string
          title: Rules Version
        classifier_id:
          type: string
          title: Classifier Id
        classifier_version:
          type: string
          title: Classifier Version
        source_row:
          anyOf:
          - type: integer
          - type: 'null'
          title: Source Row
        flags:
          additionalProperties:
            type: boolean
          type: object
          title: Flags
        primary_services:
          items:
            type: string
          type: array
          title: Primary Services
        suggestions:
          items:
            $ref: '#/components/schemas/RoutingSuggestion'
          type: array
          title: Suggestions
        excluded:
          items:
            $ref: '#/components/schemas/RoutingExclusion'
          type: array
          title: Excluded
        warnings:
          items:
            type: string
          type: array
          title: Warnings
        unresolved:
          items:
            $ref: '#/components/schemas/RoutingUnresolved'
          type: array
          title: Unresolved
      type: object
      required:
      - rules_version
      - classifier_id
      - classifier_version
      - source_row
      - flags
      - suggestions
      - excluded
      - warnings
      title: RoutingPreview
    RoutingSuggestion:
      properties:
        service:
          type: string
          title: Service
        mappings:
          items:
            $ref: '#/components/schemas/RoutingMapping'
          type: array
          title: Mappings
        primary:
          type: boolean
          title: Primary
          default: false
      type: object
      required:
      - service
      - mappings
      title: RoutingSuggestion
...
    RoutingUnresolved:
      properties:
        service:
          type: string
          title: Service
        cell:
          type: string
          title: Cell
        reason:
          type: string
          title: Reason
        source_value:
          type: string
          title: Source Value
      type: object
      required:
      - service
      - cell
      - reason
      - source_value
      title: RoutingUnresolved
...